Refresh variant stock quantity from db post-change
The stock quantity_allocated field is implemented using an F expression. This
value needs to be compared to the quantity required in a line item to ensure
stock availability.

Previously this comparison was only made once per variant per order. However
custom attributes allow mutliple order lines to contain the same variant. This
introduces a race condition whereby the comparison is made to an F expression
rather than the value of the field. Explicitly refreshing quantity_allocated
from the database mitigates this.